Upgrade to SM 60

This is the mozjs_sys smup PR.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/servo/mozjs/138)
<!-- Reviewable:end -->

+++ mozjs/testing/mozbase/mozrunner/mozrunner/devices/emulator.py
@@ -2,6 +2,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+from __future__ import absolute_import
+
 from telnetlib import Telnet
 import datetime
 import os
@@ -20,8 +22,9 @@ from ..errors import TimeoutException
 
 
 class ArchContext(object):
+
     def __init__(self, arch, context, binary=None, avd=None, extra_args=None):
-        homedir = getattr(context,'homedir', '')
+        homedir = getattr(context, 'homedir', '')
         kernel = os.path.join(homedir, 'prebuilts', 'qemu-kernel', '%s', '%s')
         sysdir = os.path.join(homedir, 'out', 'target', 'product', '%s')
         self.extra_args = []
@@ -49,6 +52,7 @@ class ArchContext(object):
 
 
 class SDCard(object):
+
     def __init__(self, emulator, size):
         self.emulator = emulator
         self.path = self.create_sdcard(size)
@@ -138,7 +142,7 @@ class BaseEmulator(Device):
 
     def _get_online_devices(self):
         return [d[0] for d in self.dm.devices() if d[1] != 'offline' if
-                    d[0].startswith('emulator')]
+                d[0].startswith('emulator')]
 
     def connect(self):
         """
@@ -160,9 +164,11 @@ class BaseEmulator(Device):
         if self.proc:
             self.proc.kill()
             self.proc = None
+            self.connected = False
 
         # Remove temporary files
-        shutil.rmtree(self.tmpdir)
+        if os.path.isdir(self.tmpdir):
+            shutil.rmtree(self.tmpdir)
 
     def _get_telnet_response(self, command=None):
         output = []
@@ -190,6 +196,7 @@ class BaseEmulator(Device):
 
 
 class Emulator(BaseEmulator):
+
     def __init__(self, app_ctx, arch, resolution=None, sdcard=None, userdata=None,
                  no_window=None, binary=None, **kwargs):
         super(Emulator, self).__init__(app_ctx, arch=arch, binary=binary, **kwargs)
@@ -216,11 +223,11 @@ class Emulator(BaseEmulator):
         """
         qemu_args = super(Emulator, self).args
         qemu_args.extend([
-                     '-kernel', self.arch.kernel,
-                     '-sysdir', self.arch.sysdir,
-                     '-data', self.userdata.name,
-                     '-initdata', self.initdata,
-                     '-wipe-data'])
+            '-kernel', self.arch.kernel,
+            '-sysdir', self.arch.sysdir,
+            '-data', self.userdata.name,
+            '-initdata', self.initdata,
+            '-wipe-data'])
         if self.no_window:
             qemu_args.append('-no-window')
         if self.sdcard:
@@ -256,7 +263,9 @@ class Emulator(BaseEmulator):
         # Remove temporary files
         self.userdata.close()
 
+
 class EmulatorAVD(BaseEmulator):
+
     def __init__(self, app_ctx, binary, avd, port=5554, **kwargs):
         super(EmulatorAVD, self).__init__(app_ctx, binary=binary, avd=avd, **kwargs)
         self.port = port
